feat: most used filter logic and style be applyed
